code cleanup per review comments
